requirements.txt
# Core Python
numpy==1.26.4

# Vector Search
faiss-cpu==1.12.0

# Embeddings
sentence-transformers==5.1.2
transformers>=4.41.0
torch>=2.0.0

# LLM Client (LiteLLM + OpenAI-compatible)
openai==2.7.1
litellm==1.79.3
httpx>=0.28.0

# Web UI
streamlit==1.39.0

# Utilities
tqdm
scikit-learn
scipy
python-dotenv
